Refactoring locale loading in play
Moving locales.ts from i18n to play directory (it is not used in pusher anymore)
Removing dynamic loading of env vars hack (because not needed anymore cause i18n is not loaded in pusher anymore)
Adding a lock to prevent loading locales twice.
